tuple: rework updates to improve code extendibility
Before the patch update was implemented as a set of operations applicable for
arrays only. It was ok until field names and JSON paths appearance, because
tuple is array on the top level.

But now there are two reasons to allow more complex updates of tuple field
internals via JSON paths:
* tuple field access by JSON path is allowed so for consistency
  JSON paths should be allowed in updates as well;
* JSON indexes are supported. JSON update should be able to change an indexed
  field without rewriting half of a tuple;
* Tarantool is going to support documents in storage so JSON
  path updates is one more step forward.

The patch prepares current update code to be extended by updates of map, so
named 'bar' updates, and multilevel updates.

The concept is to build a tree of update objects. Each updates a scalar terminal
value (leaf of that tree), or a complex object like array or map. In the latter
case these objects contain >= 2 children updates.

This organization allow to split update into multiple independent parts for each
update type.

Needed for #1261