test: join_vclock test fails under high load
Join vclock test failed on high load:

[015] Test failed! Result content mismatch:
[015] --- replication/join_vclock.result Fri Apr 12 14:49:26 2019
[015] +++ replication/join_vclock.reject Mon Apr 15 17:10:13 2019
[015] @@ -72,7 +72,7 @@
[015] ...
[015] box.space.test.index.primary:max()[1] == cnt - 1
[015] ---
[015] -- true
[015] +- false
[015] ...
[015] test_run:cmd("switch default")
[015] ---

ERRINJ_RELAY_FINAL_SLEEP stops until stop_vclock (see
src/box/relay.cc::relay_final_join() for that code and
src/box/box.cc::box_process_join() for the protocol description). So
after it unblocks a data transfer can be in a progress (during final
join or in subscribed state).

The test case checks that there are no holes at this point, so waiting
here looks inappropriate.

The real problem here is that we check amount of tuples in the space and
a content of a last tuple in a separate requests (test-run sends each
line separately to tarantool).

Close #4160