lua: return getaddrinfo() errors
Before this patch, branch when getaddrinfo() returns error codes
couldn't be reached on Mac OS, because they are greater than 0 on
Mac OS (assumption "rc < 0" in commit ea1da04d5 is incorret for
Mac OS). Add this errors into the socket and the net.box modules.
Now getaddrinfo() can return a pair of values (nil and error
message) in case of error.

Closes #4138