update: enable JSON bar updates
Bar update is an update by JSON path of a tuple internal object
with no other updates along the path. It is an optimization to do
not store each part of the path and link to it in memory. Bar
stores just string with JSON and pointer to the MessagePack
object to update.

So far bar can not have another updates along the path, this
patch introduces only non-intersected JSON path updates (that is
actually one of the most common cases, so bar is really useful
optimization).

An example. There is an update {'=', 'a.b.c[1][2][3]', 100}.
For this whole update only one object will be created:

    {
        op = '=',
        path = 'a.b.c[1][2][3]',
        new_value = 100,
        ptr_to_tuple_msgpack = 0x....,
        ...
    }

The elements 'a', 'b', 'c', '[1]', '[2]', '[3]' are not a list
of objects each occupying memory. The whole bar is stored as one
object.

Part of #1261